【调整】修改著作成果的绩效规则

+++ src/main/java/top/alvinsite/demo/model/entity/performance/BaseEntity.java
@@ -28,5 +28,5 @@ public abstract class BaseEntity {
     protected Integer SignedOrder;
 
     @ExcelColumn(value = "科研分数", col = 20)
-    protected float score;
+    protected double score;
 }
