+++ src/main/java/top/alvinsite/demo/service/performance/impl/LiteratureServiceImpl.java
@@ -39,7 +39,7 @@ public class LiteratureServiceImpl extends AbstractPerformanceService<Literature
         }
 
         // 读取计分规则
-        float score = literatureRuleService.getScore(project);
+        double score = literatureRuleService.getScore(project);
 
         // 分值分配法
         score *= scoreDistributionService.getProportion(ScoreDistributionParam.build(project, PERFORMANCE));
